Allow non-admin users via iControlRESTTokenAuth
Users that do not have the administrator role cannot use basic HTTP
authentication for BIG-IP iControl REST.  They must instead get an
authentication token that they provide with each request, as described
here:

  https://devcentral.f5.com/wiki/icontrol.authentication_with_the_f5_rest_api.ashx

The requests library provides an Auth framework: an instance of a class
derived from requests.auth.AuthBase can be attached to a
requests.Session.  This authenticator is consulted on each request and
gets to do whatever it wants to the request to make it authenticated.

For BIG-IP token auth, it will see if it already has a token.  If not, it
will attempt to go get one at mgmt/shared/authn/login.  Then the token
is stuffed into the X-F5-Auth-Token header.

BIG-IP reports a token expiration time.  We adjust it so it is
referenced to our local clock, and also expire it a little earlier (60s)
if possible.  Tokens are valid for 60 minutes by default.  Once expired,
we'll go get a new token.

This commit also adds some functional tests.  A subset can run without
requiring any new config on the BIG-IP.  To test with a non-admin user,
the BIG-IP needs one already configured.  To avoid burdening test
runners too much, these tests automatically skip unless the
"--nonadmin-username" and "--nonadmin-password" options are passed.