Fixes #119 (#123)
* Fixes #119

Problem:

_validate_suffix_collections method assumes that '~' and '/' in elements are incorrect. This leads to issues where 'name' contains '~' or '/' on purpose.

Analysis:
This patch introduces name_transform bool (with default being False), with it set to True, the 'name' element will not be validated for the presence of '~' and '/' characters. It will also convert all '/' in the name into '~'. To use this ability BIGIP needs to pass the argument 'name_transform' in the **kwargs when calling the iControl.

Tests:
unit
functional

* -added v12_1 test

* added skip ifs and added release option to testing suite

* fixing travis errors

* more jenkins fixes

* fixing no defaults for transform_name