Adds ability to set a pre-existing token in the object
Previously, the session object only allowed for the creation of new
tokens when using it. This prevented systems like Ansible from storing
tokens in cookies for "later" authentication.

This patch adds the ability to specify a token that you have previously
cached. This can be done when instantiating the session object, or
afterwards by setting the known token via the token property of the
session object