Adds new auth_provider method (#142)
This new method is required by BIG-IQ because it forces you to specify
the auth provider when you auth a user. BIG-IP also does this, but in
the past we covered that up with the "token=*" argument. This is actually
an incorrect abstraction.

Instead, this auth_provider method is the correct way. It can accept
strings such as "local" (for IQ), "tmos" (for IP), and "default" for
both IQ and IP. There is also the ability to provide custom auth
providers that can be created in IQ.